+++ src/main/java/flab/payment_system/domain/batch/compensation/CompensationItemWrite.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.batch.compensation;
 
-import flab.payment_system.domain.payment.domain.Payment;
+import flab.payment_system.domain.payment.entity.Payment;
 import jakarta.persistence.EntityManagerFactory;
 import org.springframework.batch.item.database.JpaItemWriter;
 import org.springframework.context.annotation.Bean;
