+++ src/main/java/flab/payment_system/domain/payment/entity/kakao/ApprovedCancelAmount.java
@@ -1,4 +1,4 @@
-package flab.payment_system.domain.payment.domain.kakao;
+package flab.payment_system.domain.payment.entity.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -6,11 +6,11 @@ import lombok.Getter;
 
 @Getter
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class CancelAvailableAmount {
+public class ApprovedCancelAmount {
 
-	Integer total;
-	Integer taxFree;
-	Integer vat;
-	Integer point;
-	Integer discount;
+	private Integer total;
+	private Integer taxFree;
+	private Integer vat;
+	private Integer point;
+	private Integer discount;
 }
