+++ src/main/java/flab/payment_system/domain/payment/entity/kakao/CanceledAmount.java
@@ -1,24 +1,16 @@
-package flab.payment_system.domain.payment.domain.kakao;
+package flab.payment_system.domain.payment.entity.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import jakarta.persistence.Column;
-import jakarta.persistence.Embeddable;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-@Embeddable
-public class Amount {
+public class CanceledAmount {
 
 	private Integer total;
-	@Column(name = "tax_free")
 	private Integer taxFree;
-	private Integer tax;
+	private Integer vat;
 	private Integer point;
 	private Integer discount;
-	@Column(name = "green_deposit")
-	private Integer greenDeposit;
 }
