+++ src/main/java/flab/payment_system/domain/payment/entity/toss/CashReceipt.java
@@ -1,4 +1,4 @@
-package flab.payment_system.domain.payment.domain.toss;
+package flab.payment_system.domain.payment.entity.toss;
 
 import lombok.Getter;
 import lombok.Setter;
