+++ src/main/java/flab/payment_system/domain/payment/entity/toss/Checkout.java
@@ -1,11 +1,10 @@
-package flab.payment_system.domain.payment.domain.toss;
+package flab.payment_system.domain.payment.entity.toss;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class Receipt {
-
+public class Checkout {
 	private String url;
 }
