+++ src/main/java/flab/payment_system/domain/payment/entity/toss/Failure.java
@@ -1,4 +1,4 @@
-package flab.payment_system.domain.payment.domain.toss;
+package flab.payment_system.domain.payment.entity.toss;
 
 import lombok.Getter;
 import lombok.Setter;
