+++ src/main/java/flab/payment_system/domain/payment/entity/toss/Receipt.java
@@ -1,10 +1,11 @@
-package flab.payment_system.domain.payment.domain.toss;
+package flab.payment_system.domain.payment.entity.toss;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class Checkout {
+public class Receipt {
+
 	private String url;
 }
