+++ src/main/java/flab/payment_system/domain/payment/repository/PaymentRepository.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.payment.repository;
 
-import flab.payment_system.domain.payment.domain.Payment;
+import flab.payment_system.domain.payment.entity.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
