+++ src/main/java/flab/payment_system/domain/payment/repository/kakao/KakaoPaymentRepository.java
@@ -1,6 +1,6 @@
 package flab.payment_system.domain.payment.repository.kakao;
 
-import flab.payment_system.domain.payment.domain.kakao.KakaoPayment;
+import flab.payment_system.domain.payment.entity.kakao.KakaoPayment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface KakaoPaymentRepository extends JpaRepository<KakaoPayment, Long>,
