+++ src/main/java/flab/payment_system/domain/payment/response/kakao/PaymentKakaoApprovalDtoImpl.java
@@ -2,8 +2,8 @@ package flab.payment_system.domain.payment.response.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import flab.payment_system.domain.payment.domain.kakao.Amount;
-import flab.payment_system.domain.payment.domain.kakao.CardInfo;
+import flab.payment_system.domain.payment.entity.kakao.Amount;
+import flab.payment_system.domain.payment.entity.kakao.CardInfo;
 import flab.payment_system.domain.payment.response.PaymentApprovalDto;
 import lombok.Getter;
 import lombok.Setter;
