+++ src/main/java/flab/payment_system/domain/payment/response/kakao/PaymentKakaoOrderDetailDtoImpl.java
@@ -2,10 +2,10 @@ package flab.payment_system.domain.payment.response.kakao;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import flab.payment_system.domain.payment.domain.kakao.Amount;
-import flab.payment_system.domain.payment.domain.kakao.CancelAvailableAmount;
-import flab.payment_system.domain.payment.domain.kakao.CanceledAmount;
-import flab.payment_system.domain.payment.domain.kakao.PaymentActionDetails;
+import flab.payment_system.domain.payment.entity.kakao.Amount;
+import flab.payment_system.domain.payment.entity.kakao.CancelAvailableAmount;
+import flab.payment_system.domain.payment.entity.kakao.CanceledAmount;
+import flab.payment_system.domain.payment.entity.kakao.PaymentActionDetails;
 import flab.payment_system.domain.payment.response.PaymentOrderDetailDto;
 import lombok.Getter;
