+++ src/main/java/flab/payment_system/domain/payment/response/toss/PaymentTossDtoImpl.java
@@ -1,18 +1,18 @@
 package flab.payment_system.domain.payment.response.toss;
 
-import flab.payment_system.domain.payment.domain.toss.Cancels;
-import flab.payment_system.domain.payment.domain.toss.Card;
-import flab.payment_system.domain.payment.domain.toss.CashReceipt;
-import flab.payment_system.domain.payment.domain.toss.CashReceipts;
-import flab.payment_system.domain.payment.domain.toss.Checkout;
-import flab.payment_system.domain.payment.domain.toss.Discount;
-import flab.payment_system.domain.payment.domain.toss.EasyPay;
-import flab.payment_system.domain.payment.domain.toss.Failure;
-import flab.payment_system.domain.payment.domain.toss.GiftCertificate;
-import flab.payment_system.domain.payment.domain.toss.MobilePhone;
-import flab.payment_system.domain.payment.domain.toss.Receipt;
-import flab.payment_system.domain.payment.domain.toss.Transfer;
-import flab.payment_system.domain.payment.domain.toss.VirtualAccount;
+import flab.payment_system.domain.payment.entity.toss.Cancels;
+import flab.payment_system.domain.payment.entity.toss.Card;
+import flab.payment_system.domain.payment.entity.toss.CashReceipt;
+import flab.payment_system.domain.payment.entity.toss.CashReceipts;
+import flab.payment_system.domain.payment.entity.toss.Checkout;
+import flab.payment_system.domain.payment.entity.toss.Discount;
+import flab.payment_system.domain.payment.entity.toss.EasyPay;
+import flab.payment_system.domain.payment.entity.toss.Failure;
+import flab.payment_system.domain.payment.entity.toss.GiftCertificate;
+import flab.payment_system.domain.payment.entity.toss.MobilePhone;
+import flab.payment_system.domain.payment.entity.toss.Receipt;
+import flab.payment_system.domain.payment.entity.toss.Transfer;
+import flab.payment_system.domain.payment.entity.toss.VirtualAccount;
 import flab.payment_system.domain.payment.response.PaymentApprovalDto;
 import flab.payment_system.domain.payment.response.PaymentCancelDto;
 import flab.payment_system.domain.payment.response.PaymentOrderDetailDto;
